pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zava-signal-analyst"
version = "1.0.0"
description = "Zava Signal Analyst — A2A Protocol agent for Copilot Studio"
requires-python = ">=3.11"
dependencies = [
    # AI / Agent framework
    "agent-framework-core==1.0.0rc1",
    "agent-framework-azure-ai==1.0.0rc1",
    "azure-identity",

    # OpenTelemetry semantic conventions (LLM attributes for agent-framework)
    "opentelemetry-semantic-conventions-ai==0.4.13",

    # HTTP server
    "aiohttp>=3.9.0",

    # Data models & env
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",

    # HTTP clients (tools)
    "httpx>=0.27.0",

    # Browser automation
    "playwright>=1.40.0",

    # PDF reports
    "reportlab>=4.0",

    # Word document generation
    "python-docx>=1.1.0",

    # Azure Blob Storage (report persistence)
    "azure-storage-blob>=12.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-aiohttp>=1.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
